repo setup (#1)
* repo setup

* fixes

* feat: made the prisma package

* Update apps/web/src/components/Admin/Game/WaitScreen.tsx


* fixes

* fixes

* fixes

* fixes

---------

+++ apps/web/src/actions/DeleteQuizAction.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { auth } from "@/utils/auth";
-import { prisma } from "@/utils/prisma";
+import { prisma } from "@buzrr/prisma";
 import { redirect } from "next/navigation";
 import { revalidatePath } from "next/cache";
 
@@ -10,10 +10,15 @@ async function dltQuiz(quizId: string) {
     const session = await auth();
     if (!session || !session.user) redirect("/api/auth/signin");
 
+    const quiz = await prisma.quiz.findFirst({
+      where: { id: quizId, userId: session.user.id },
+    });
+    if (!quiz) {
+      return { error: "Unauthorized or quiz not found" };
+    }
+
     await prisma.quiz.delete({
-      where: {
-        id: quizId,
-      },
+      where: { id: quizId },
     });
 
     revalidatePath("/admin", "page");
